Automatically determine FlightGear version under Windows
Under Windows the .bat scripts are used to start a FlightGear
visualisation. In these scripts the FlightGear version was
hard coded so that they did not work for different versions.
The problem was solved by adding the .bat script
fg_install_searcher.bat. This script searches for the
FlightGear installation directory. If multiple FlightGear
installations are found, the user is asked to select the
desired version.