Fix FlightGear startup under Windows
The .bat scripts to start FlightGear under Windows only worked if they were executed from the same drive where FlightGear is installed (usually C:\). This update fixes the problem of FlightGear not loading if the start scripts are on a different drive. In addition, the FlightGear terrasync option has been added to visualise the environment as is already the case in the bash scripts.